Fixata requestUri lato server e cambiato le std::cout con le qDebug

+++ CLIENT/connection_to_server.h
@@ -28,7 +28,7 @@ public:
     std::string requestGetNickname(int userId);
     std::string requestGetUsername(int userId);
     std::string requestIconId(int userId);
-    std::string requestUri(long docId);
+    std::string requestUri(int docId);
     std::string requestDocDatoUri(QString uri);
     std::shared_ptr<QMap<QString, int>> getKnownDocuments(int userId);
 private slots:
